emacs : fix circadian not working with the modus themes
For some reason, loading a modus-theme with a second
argument (load-theme theme t) instead of (load-theme theme) causes the
loading of the theme to fail.
Function circadian-enable-theme is therefore modified in order to load
the theme in an appropriate way. Emacs may ask if the theme is safe to
be executed as a downside, but putting the theme in the
custom-safe-themes is only a one time inconvenience.